requirements.txt
numpy>=1.21.0
Pillow>=8.0.0
scikit-image>=0.18.0
matplotlib>=3.4.0
gurobipy>=9.1.0
pandas>=1.3.0
scipy>=1.7.0
